requirements.txt
flask
flask-cors
python-docx
openai==1.0.0
gunicorn
